Convert ndkports into a gradle plugin.
This lets us use a lot of more the built in gradle behavior for things
like maven publishing and zip tasks, as well as allows us to split
individual ports out into their own projects more easily.

It also has the handy side-effect of relying on prefab to make use of
any packages that are depended on by the ports, so we can be more
certain that the openssl AAR is working because the curl port needs to
use it.

Updates to Gradle 6.8 because Gradle 6.7 is still on Kotlin 1.3 and
both the kotlin-script plugin and java-gradle-plugin dependendencies
force 1.3 onto our class path, which emits a bunch of warnings saying
that we're probably broken.
https://youtrack.jetbrains.com/issue/KT-3801

Test: ./gradlew build
Bug: None
Change-Id: Ia4da7a6a50a5f6fbfa51441497347ad49178a2b0